backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
yt-dlp>=2024.12.0
groq>=0.11.0
google-generativeai>=0.8.0
supabase>=2.4.0
gradio>=4.44.0
